requirements.txt
# Core dependencies
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
SQLAlchemy==2.0.22
pymysql==1.1.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0

# Logging and monitoring
structlog==23.2.0

# Utilities
tenacity==8.2.3  # For retries
cryptography>=41.0.3  # Required by pymysql for secure connections